RestController

+++ src/main/java/com/digis01/IHernandezProgramacionNCapas/DAO/UsuarioJPADAOImplementation.java
@@ -1,5 +1,6 @@
 package com.digis01.IHernandezProgramacionNCapas.DAO;
 
+import com.digis01.IHernandezProgramacionNCapas.JPA.Direccion;
 import com.digis01.IHernandezProgramacionNCapas.JPA.Result;
 import com.digis01.IHernandezProgramacionNCapas.JPA.Usuario;
 import jakarta.persistence.EntityManager;
@@ -24,12 +25,14 @@ public class UsuarioJPADAOImplementation implements IUsuarioJPADAO
             result.object = queryUsuario.getResultList();
             
             result.correct = true;
+            result.status = 200;
         } 
         catch (Exception ex) 
         {
             result.correct = false;
             result.errorMessage = ex.getLocalizedMessage();
             result.ex = ex;
+            result.status = 500;
         }
         return result;
     }
@@ -42,9 +45,18 @@ public class UsuarioJPADAOImplementation implements IUsuarioJPADAO
         try 
         {
             Usuario usuarioJPA = entityManager.find(Usuario.class, idUsuario);
-            result.object = usuarioJPA;
             
-            result.correct = true;
+            if (usuarioJPA != null) 
+            {
+                result.object = usuarioJPA;
+                result.correct = true;
+                result.status = 200;
+            } else
+            {
+                result.errorMessage = "Usuario no encontrado, Id incorrecto";
+                result.status = 400;
+            }
+            
         } catch (Exception ex) 
         {
             result.correct = false;
@@ -55,8 +67,8 @@ public class UsuarioJPADAOImplementation implements IUsuarioJPADAO
         return result;
     }
 
-    @Override
     @Transactional
+    @Override
     public Result Add(Usuario usuario) 
     {
         Result result = new Result();
@@ -64,13 +76,22 @@ public class UsuarioJPADAOImplementation implements IUsuarioJPADAO
         try 
         {
             entityManager.persist(usuario);
+            result.object = usuario;
+            
+            Direccion direccion = usuario.Direcciones.get(0);
+            direccion.Usuario = usuario;
+            direccion.Usuario.setIdUsuario(usuario.getIdUsuario());
+            entityManager.persist(direccion);
             
             result.correct = true;
+            result.status = 200;
+            
         } catch (Exception ex) 
         {
             result.correct = false;
             result.errorMessage = ex.getLocalizedMessage();
             result.ex = ex;
+            result.status = 500;
         }
         
         return result;
@@ -85,9 +106,18 @@ public class UsuarioJPADAOImplementation implements IUsuarioJPADAO
         try 
         {
             Usuario usuario = entityManager.find(Usuario.class, idUsuario);
-            entityManager.remove(usuario);
-            result.correct = true;
             
+            if (usuario != null) 
+            {
+                entityManager.remove(usuario);
+                result.correct = true;
+                result.status = 200;
+            } else
+            {
+                result.errorMessage = "Usuario no encontrado, Id incorrecto";
+                result.status = 400;
+            }
+           
         } catch (Exception ex) 
         {
             result.correct = false;
@@ -107,18 +137,73 @@ public class UsuarioJPADAOImplementation implements IUsuarioJPADAO
         try 
         {
             Usuario usuarioBD = entityManager.find(Usuario.class, usuario.getIdUsuario());
-            Usuario usuarioJPA = new Usuario();
-            usuarioJPA.Direcciones = usuarioBD.Direcciones;
-            entityManager.merge(usuarioJPA);
             
-            result.correct = true;
-            
+            if(usuarioBD != null)
+            {
+                usuarioBD.setStatus(usuario.getStatus());
+                usuarioBD.setImagen(usuario.getImagen());
+                usuarioBD.setUsername(usuario.getUsername());
+                usuarioBD.setNombre(usuario.getNombre());
+                usuarioBD.setApellidoPaterno(usuario.getApellidoPaterno());
+                usuarioBD.setApellidoMaterno(usuario.getApellidoMaterno());
+                usuarioBD.setFechaNacimiento(usuario.getFechaNacimiento());
+                usuarioBD.setSexo(usuario.getSexo());
+                usuarioBD.setCurp(usuario.getCurp());
+                usuarioBD.setEmail(usuario.getEmail());
+                usuarioBD.setPassword(usuario.getPassword());
+                usuarioBD.setTelefono(usuario.getTelefono());
+                usuarioBD.setCelular(usuario.getCelular());
+                usuarioBD.Rol = usuario.Rol;
+                
+                entityManager.merge(usuarioBD);
+                result.correct = true;
+                result.object = usuarioBD;
+                result.status = 200;
+            } else
+            {
+                result.errorMessage = "Usuario no encontrado, Id incorrecto";
+                result.status = 400;
+            }    
         } catch (Exception ex) 
         {
             result.correct = false;
             result.errorMessage = ex.getLocalizedMessage();
             result.ex = ex;
         }       
+        return result;
+    }
+    
+    @Transactional
+    @Override
+    public Result UpdateStatus(Usuario usuario)
+    {
+        Result result = new Result();
+        
+        try 
+        {
+            Usuario usuarioBD = entityManager.find(Usuario.class, usuario.getIdUsuario());
+            
+            if(usuarioBD != null)
+            {
+                usuarioBD.setStatus(usuario.getStatus());
+                entityManager.merge(usuarioBD);
+                
+                result.object = usuarioBD;
+                result.correct = true;
+                result.status = 200;
+            } else
+            {
+                result.errorMessage = "Usuario no encontrado, Id incorrecto";
+                result.status = 400;
+            }
+            
+        } catch (Exception ex) 
+        {
+            result.correct = false;
+            result.errorMessage = ex.getLocalizedMessage();
+            result.ex = ex;
+        }
+        
         return result;
     }
 }
